Add WebVTT Bottom Aligment and center for horiz.
WebVTT aligment algorithms consider bottom aligments as
default so that line: is not added (or read from existing VTT)
for bottom aligned.

These aligments were added so they'd be both exported and imported when reading/writting a VTT file.

So far, bottom aligment is hard coded, at 85% for one
line, 79% for two lines. In the future I will shift this to a
dialog setting where you can set how low middle and
bottom aligments are and a modifier for each extra line.

This was done because bottom aligment doesn't seem
to be the default for all programs, for example
telestream treats no line parameter as a top aligment.

Also, for all central aligments (an 2, 5 and 8) I've added
align:middle as a simple possition:50% might cause
lines to start on the middle and run to the right.